[Class] Changing file bor for format class
